refactor:change path of fs-tasks [list,read,rename]
